refactor : String carName -> CarName carName 문자열과 원시값 포장처리
2021. 10. 08.
- 자동차이름을 CarName으로 포장
- CarNameTest 추가
- 관련 로직 수정

Resolver : loop-study
See also : NEXTSTEP
