+++ README.md
@@ -78,11 +78,15 @@ honux : -----
 
 ## 구현 기능 목록
 - [x] Car : 자동차
-  - [x] 이름을 갖는다.
+  - [x] CarName(이름)을 갖는다.
   - [x] 거리를 갖는다.
   - [x] 0 ~ 9 사이의 Random 값으로 전진 여부를 결정한다.
   - [x] 4 이상이면 한칸 전진한다.
   - [x] 3 이하는 멈춘다.
+
+- [x] CarName : 자동차이름
+  - [x] 이름은 1~5글자로 제한한다.
+  - [x] 잘못된 이름은 예외처리 한다.
 
 - [x] Cars : 자동차 여러대
   - [x] 자동차 여러대를 갖는다.
