+++ src/main/java/racinggame/domain/Car.java
@@ -1,25 +1,19 @@
 package racinggame.domain;
 
 import racinggame.strategy.Strategy;
-import racinggame.utils.ValidationUtils;
-
-import java.util.Objects;
 
 public class Car {
-    private static final String ERROR_CAR_NAME_MESSAGE = "자동차 이름이 잘못되었습니다.";
-
-    private String carName;
+    private CarName carName;
     private int distance;
     private Strategy strategy;
 
     public Car(String carName, Strategy strategy) {
-        validationCarName(carName);
-        this.carName = carName;
+        this.carName = new CarName(carName);
         this.strategy = strategy;
         this.distance = 0;
     }
 
-    private Car(String carName, Strategy strategy, Integer distance) {
+    private Car(CarName carName, Strategy strategy, Integer distance) {
         this.carName = carName;
         this.strategy = strategy;
         this.distance = distance;
@@ -33,17 +27,11 @@ public class Car {
     }
 
     public String getCarName() {
-        return carName;
+        return carName.getName();
     }
 
     public int getDistance() {
         return distance;
-    }
-
-    private void validationCarName(String carName) {
-        if (!ValidationUtils.validCarName(carName)) {
-            throw new IllegalArgumentException(ERROR_CAR_NAME_MESSAGE);
-        }
     }
 
     public boolean isMaxDistance(int distance) {
